refactor | s&t : export all through sal-3dlib
