+++ crates/sal-3dlib-core/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "sal-3dlib-core"
+version = "0.0.1"
+edition = "2021"
