+++ crates/sal-3dlib-core/src/bound.rs
