+++ crates/sal-3dlib-core/src/gmath/vector.rs
