+++ crates/sal-3dlib-core/src/lib.rs
@@ -1,0 +1,8 @@
+//! Abstact definition of objects and operations over the CAD kernel.
+//! It is assumed that the kernel is universal, but it implements all the necessary traits.
+//
+pub mod bound;
+pub mod gmath;
+pub mod ops;
+pub mod props;
+pub mod topology;
