+++ crates/sal-3dlib-core/src/ops.rs
