+++ crates/sal-3dlib-core/src/ops/boolean.rs
