+++ crates/sal-3dlib-core/src/props.rs
