+++ crates/sal-3dlib-core/src/topology.rs
