+++ crates/sal-3dlib-core/src/topology/shape/solid.rs
