+++ crates/sal-3dlib/tests/hull_shell.step
