+++ crates/sal-occt-rs/src/ops.rs
@@ -1,3 +1,3 @@
-//! Implementation of [sal_3dlib::ops].
+//! Implementation of [sal_3dlib_core::ops].
 //
 pub mod boolean;
