+++ crates/sal-occt-rs/src/ops/boolean.rs
@@ -1,14 +1,14 @@
-//! Implementation of [sal_3dlib::ops::boolean].
+//! Implementation of [sal_3dlib_core::ops::boolean].
 //
 pub mod volume;
 //
-use sal_3dlib::ops::boolean;
+use sal_3dlib_core::ops::boolean;
 ///
-/// See [sal_3dlib::ops::boolean::OpConf] for details.
+/// See [sal_3dlib_core::ops::boolean::OpConf] for details.
 #[derive(Clone, Copy)]
 pub struct OpConf {
     ///
-    /// See [sal_3dlib::ops::boolean::OpConf::parallel] for details.
+    /// See [sal_3dlib_core::ops::boolean::OpConf::parallel] for details.
     pub parallel: bool,
 }
 //
