+++ crates/sal-occt-rs/src/topology.rs
@@ -1,3 +1,3 @@
-//! Implementation of [sal_3dlib::topology].
+//! Inner structures used for [sal_3dlib_core::topology::Shape].
 //
 pub mod shape;
